Add search and cache services.
