+++ src/db/redis.py
@@ -1,8 +1,22 @@
+from abc import ABC, abstractmethod
+
 from redis.asyncio import Redis
 
-redis: Redis = Redis()
+
+class AsyncCacheInterface(ABC):
+    @abstractmethod
+    async def get_cache_connection(self):
+        pass
 
 
-# Функция понадобится при внедрении зависимостей
-async def get_redis() -> Redis:
-    return redis
+class AsyncCacheImplementation(AsyncCacheInterface):
+    async def get_cache_connection(self):
+        return Redis()
+
+
+async def main():
+    cache_implementation = AsyncCacheImplementation()
+    conn = await cache_implementation.get_cache_connection()
+    return conn
+
+redis = AsyncCacheImplementation()
